Modified api.js
